idea col list container
